fix: FavoriteActivity removed
The reviewer said "We need to show the list of favourite movies by updating the main view and we do not need a separate activity for this.", so we removed the activity. Now, all the favorite movies are displayed in the main ctivity based on the shared preferences.

Resolves:NA
See also:NA